Reorganize ThemeContext
