+++ .storybook/main.ts
@@ -13,6 +13,7 @@ const config: StorybookConfig = {
         name: '@storybook/react-vite',
         options: {},
     },
+    staticDirs: ['../public'],
 };
 // noinspection JSUnusedGlobalSymbols
 export default config;
