+++ src/main.style.tsx
@@ -2,8 +2,8 @@ import { useContext } from 'react';
 
 import { createGlobalStyle } from 'styled-components';
 
-import { darkColors, lightColors } from '@/shared/constants/colors.ts';
-import { ThemeContext } from '@/shared/contexts/Theme.tsx';
+import { darkColors, lightColors } from '@/shared/constants/colors';
+import { ThemeContext } from '@/shared/contexts/ThemeContext';
 
 // noinspection CssUnknownTarget
 const CommonGlobalStyle = createGlobalStyle`
