+++ src/shared/components/Text/index.ts
@@ -1,0 +1,3 @@
+import Text from '@/shared/components/Text/Text';
+
+export default Text;
